Improve memory use efficiently by 40+%
This is done by adjusting the rump kernel memory allocation hypercall
to directly use the page allocator for page-sized allocation (the very
common case).  This in turn avoids having to allocate an extra page for
the malloc header, which in turn makes rump_hypermalloc(PAGE_SIZE)
consume only one page of memory instead of two.