rumpxendev: Provide /dev/xen/xenbus
This is a new rumpkernel component which provides an interface to the
minios xenbus driver.  It presents as /dev/xen/xenbus which speaks the
standard xenstore protocol and can be used in the normal way by all
the tools (including with poll or select).

Internally, it arranges to use the xenbus driver's ability to
multiplex requests, responses and watch events from multiple users.
So it can be used in parallel with all existing rump kernel
functionality.  The device can also perhaps be opened more than once.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>