Speed up app-configure by a factor of many.
This is done by not linking in the full rump kernel at the configure
stage.  A cross-configure configure script will only do compile tests and
it does not really care about what is present in the kernel.  Since all
of the "libc" bits are in librump (apart from ones we handle directly on
top of minios), just linking in librump at the configure stage is enough.