Localize all non-public Mini-OS symbols
Pass 3 of X in Mini-OS namespace cleanup.

We define a set of prefixes in the Makefile for the symbol namespaces we
wish to keep. Then, when linking minios.o we use objcopy to localize all
other symbols. Note the sole exception of the arch-specific startup file
(x86_64.o) as this is used as the linker %startfile.

Signed-off-by: Martin Lucina <martin@lucina.net>