Sketch out new interfaces for bootstrap/running programs.
The system is "booted" using rumprun_boot().  The opposite of
rumprun_boot() is rumprun_reboot() (logically?)

Between rumprun_boot() and rumprun_reboot() individual applications may
be launched with rumprun() (async) and waited for with rumprun_wait().

Of course, bits for supporting >1 program are still missing, but at
least the sketch is there.