First attempt at adding -n(etspec) support for qemu.
This commit also splits parse_netspec() into a backend-independent parser
and a backend-dependent store function.  Currently, the store function
is implemented for xenstore and json.

There are still some problems.  -n implicitly creates the networking
interfaces, which means that we don't necessarily get the type of
interface we want, e.g. if we want to use a backing /dev/tap.  But
all in good time ...