Rewrite the scheduler core for up to >50% better performance.
The normal case for a rump kernel is that we have 20+ unrunnable threads
and a few hot runnable ones.  The previous scheduler essentially went
over all of the unrunnable threads every time it was looking for a thread
to schedule resulting in huge amounts of unnecessary cycles in every
bmk_sched() call.  The new scheduler uses multiple queues for a O(1)
scheduling algorithm.