Actually, default to no rump kernel memory upper limit except x/2
The idea of hardcoding a specific limit was that the rump kernel wouldn't
eat too much memory as caches, but any arbitrary hardcoded upper limit
just causes confusion.  We saw that with trying to make a big tmpfs,
giving the guest lots of memory, and still having to wonder wtf things
are running out of memory.

So, I'll the memory thing properly some day, but for now just default
to x/2.