Always run page allocator on "virtual memory" (or just "memory").
The page bitmap is its own "address space", which is va offset by the
address of the first page loaded.  However, there is no assumption
anymore that this in any way maps to the physical memory space anymore.

TODO: convert Xen page mallocator to use the same scheme and unify
page allocators