hw: Fix conditions in bmk_cpu_block and simplify common codepath
While removing the while loop in bmk_cpu_block(), the wrong condition
for running into the main codepath crept in. This was manifesting as
clock running slow on x86_64.

This commit fixes the conditions and cleans up the codepath to make the
control flow for the common case easier to follow.