Build libs into objdir.
The grand implication of this commit is that now it's possible to
build both hw and xen from the same source tree (which is useful
mostly for me so that I can easily test things when developing).
nb. different target cpus are still not supported from the same
source tree.