Also add $RUMPRUN_CC to config.
A build-specific cc is needed because cc includes the lib path
in the binary for baking.  After this change it's possible to use
config variables alone to build at least a simple unikernel for
testing (i.e. setting PATH to match the build dest is not required).

XXX: value is hardcoded to gcc, but that's not a practical problem as
long as support only gcc.