genfs: pass data files to gcc via symbolic link
The filenames are interpreted by gcc, both in if gcc should
compile the files (which we don't want it, since they're data)
and also as how to transform the filename to the symbol name.
Using a known-good filename (the symlink) takes care of both issues.

problems were reported by @davedoesdev