Treat routed and non-routed interrupts separately.
For systems with routed interrupts, we don't know exactly
how they're routed (since we don't parse acpi tables or
mpbios or any of that rubbish).  However, for bog-standard
ISA interrupts we do know where they'll appear.  So avoid
calling the non-routed interrupt handler in case of a routed
interrupt and vice versa.