Search existing value of XEN_HEADERS for Xen headers
This allows a caller to specify a specific location for the Xen
headers.  For example, from a version of Xen just built in a
neighbouring directory.

(Without this it is not possible to build against a specific Xen
without being able to overwrite system header areas.)

If XEN_HEADERS is unset, or not suitable, use the existing search
strategy.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>