librumpxen_xendev: Rename applicable `d' variables to `du'
This used to be a reference to the shared device context structure.
Now there are three structures.  Now there are three structures.
Change all the variables referring to what is now struct
rumpxenbus_data_user to `du'.  This completes the elimination of the
conventional variable name `d'.  No functional change.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>